requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.95.2
flake8==6.0.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
kombu==5.2.4
mccabe==0.7.0
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycodestyle==2.10.0
pydantic==1.10.7
pydub==0.25.1
pyflakes==3.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
redis==4.5.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
